backend/requirements.txt
django
djangorestframework
django-cors-headers
django-storages
django-storages
boto3
pandas
openpyxl
gunicorn
dj_database_url
psycopg2-binary
python-dotenv
numpy
openpyxl
openai